+++ core/src/main/java/org/opensearch/remote/metadata/client/DeleteDataObjectRequest.java
@@ -8,10 +8,13 @@
  */
 package org.opensearch.remote.metadata.client;
 
+import org.opensearch.action.support.WriteRequest.RefreshPolicy;
+import org.opensearch.common.unit.TimeValue;
+
 /**
  * A class abstracting an OpenSearch DeleteRequest
  */
-public class DeleteDataObjectRequest extends WriteDataObjectRequest {
+public class DeleteDataObjectRequest extends WriteDataObjectRequest<DeleteDataObjectRequest> {
 
     /**
      * Instantiate this request with an index and id.
@@ -22,9 +25,19 @@ public class DeleteDataObjectRequest extends WriteDataObjectRequest {
      * @param tenantId the tenant id
      * @param ifSeqNo the sequence number to match or null if not required
      * @param ifPrimaryTerm the primary term to match or null if not required
+     * @param refreshPolicy when should the written data be refreshed. May not be applicable on all clients. Defaults to {@code IMMEDIATE}.
+     * @param timeout A timeout to wait if the index operation can't be performed immediately. May not be applicable on all clients. Defaults to {@code 1m}.
      */
-    public DeleteDataObjectRequest(String index, String id, String tenantId, Long ifSeqNo, Long ifPrimaryTerm) {
-        super(index, id, tenantId, ifSeqNo, ifPrimaryTerm, false);
+    public DeleteDataObjectRequest(
+        String index,
+        String id,
+        String tenantId,
+        Long ifSeqNo,
+        Long ifPrimaryTerm,
+        RefreshPolicy refreshPolicy,
+        TimeValue timeout
+    ) {
+        super(index, id, tenantId, ifSeqNo, ifPrimaryTerm, refreshPolicy, timeout, false);
     }
 
     /**
@@ -46,7 +59,15 @@ public class DeleteDataObjectRequest extends WriteDataObjectRequest {
          */
         public DeleteDataObjectRequest build() {
             WriteDataObjectRequest.validateSeqNoAndPrimaryTerm(this.ifSeqNo, this.ifPrimaryTerm, false);
-            return new DeleteDataObjectRequest(this.index, this.id, this.tenantId, this.ifSeqNo, this.ifPrimaryTerm);
+            return new DeleteDataObjectRequest(
+                this.index,
+                this.id,
+                this.tenantId,
+                this.ifSeqNo,
+                this.ifPrimaryTerm,
+                this.refreshPolicy,
+                this.timeout
+            );
         }
     }
 }